pyproject.toml
[tool.poetry]
name = "web-quotation"
version = "0.1.0"
description = ""
authors = ["YPYP333YPYP <ypyp333ypyp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
sqlalchemy = "^2.0.28"
pydantic = "^2.6.3"
alembic = "^1.13.1"
pydantic-settings ="^2.2.1"
starlette = "^0.36.3"
pandas = "^2.2.1"
aiomysql = "^0.2.0"
asyncio = "^3.4.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
asyncpg = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
